% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STPoissonSS-package.R
\docType{package}
\name{STPoissonSS-package}
\alias{STPoissonSS}
\alias{STPoissonSS-package}
\title{STPoissonSS: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Projeto-CNPq-Clima/STPoissonSS}
  \item Report bugs at \url{https://github.com/Projeto-CNPq-Clima/STPoissonSS/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
